[naming-convention] Update property schema to accept objects. Add suffix rule
To make this rule more flexible, it needs to be able to take an object as well as format style string
The config is normalised into an object before passing into the validation function.

Note that we can no longer rely on the default enum check because it also tries
to validate that an object is part of the enum. Therefore, the validation function needs to check
the format enums now.

Add suffix rule