Implement acceptedIdNames, acceptedIdTypes logic for strict-id-in-types
For a field to be valid, it must have exactly one field
that matches one of the accepted field name and one
of the accepted types.

Since we are being strict, there should only be one unique
identifier field. Note that some clients like Apollo allows
a combination of key fields to be used for caching. We can
add this feature as a separate PR if needed.

The default acceptedIdNames is ['id'] and acceptedIdTypes
is ['ID'] because most GraphQL cients use 'id' as the
field used for caching.