Implement exceptions logic for strict-id-in-types
There are certain output types that do not have natural unique
identifier e.g. Result, Payload, Pagination, etc.

strict-id-in-types has `exceptions` setting which will ignore a node
if it matches ONE of any of its sub config.
For example, `exceptions.suffixes` declares that if the node has
a name that has any of the provided suffix, it should skip the rule
check.