fix: find graphqlrc files relative to linted file
Before this change, `loadConfigSync` from graphql-config would use `process.cwd` to find any .graphqlrc files.

With a project like this:
```
project/
- packages/
  - app/
    - schema.graphql
    - .eslintrc
    - .graphqlrc
    - documents/
      - operation.graphql
```

the VSCode Eslint plugin throws an error about not being able to find the schema.graphql file when linting the operation.graphql.

This change passes in the [`rootDir` option](https://graphql-config.com/load-config#rootdir) to `loadConfigSync` using the directory of the linted file. This will allow the .graphqlrc file to live in subfolders of the open project.